% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Performance_metrics.R
\name{AAVC}
\alias{AAVC}
\title{Average Annual Variability in Yield over the first 30 projection years (a performance metrics function of class PM)}
\usage{
AAVC(MSE, pp = 1)
}
\arguments{
\item{MSE}{An object of class MSE}
}
\value{
a matrix with n Management procedures (MSE@nMP) and nsim (MSE@nsim) columns from \code{MSE}
}
\description{
Average Annual Variability in Yield over the first 30 projection years (a performance metrics function of class PM)
}
\examples{
loadABT()
AAVC(MSE_example)
}
