% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_documentation.R
\docType{data}
\name{Allocation}
\alias{Allocation}
\title{Default 2020 Allocation (a matrix area x fleet)}
\description{
Default 2020 Allocation (a matrix area x fleet)
}
\author{
Tom Carruthers \email{t.carruthers@oceans.ubc.ca}
}
\keyword{data}
