% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{areaplot}
\alias{areaplot}
\title{Plot area definitions of an OMI, OM or MSE object}
\usage{
areaplot(.Object, xadj = NA, yadj = NA)
}
\arguments{
\item{.Object}{an object of class OMI, OM or MSE}
}
\value{
a map of the areas of \code{classy}
}
\description{
Plot area definitions of an OMI, OM or MSE object
}
\examples{
loadABT()
areaplot(OM_1)
areaplot(OMI_1)
areaplot(MSE_example)
}
