% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/External.R
\name{avail}
\alias{avail}
\title{Generic class finder}
\usage{
avail(classy)
}
\arguments{
\item{classy}{An character string representing an R class}
}
\value{
a vector of objects of class\code{classy}
}
\description{
Generic class finder
}
\examples{
avail('data.frame')
avail('matrix')
avail('ABT_PM')
}
