% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Performance_metrics.R
\name{AvC30}
\alias{AvC30}
\title{Mean  catches over first 30 projected years (a performance metrics function of class PM)}
\usage{
AvC30(MSE, pp = 1)
}
\arguments{
\item{MSE}{An object of class MSE}
}
\value{
a matrix with n Management procedures (MSE@nMP) and nsim (MSE@nsim) columns from \code{MSE}
}
\description{
Mean  catches over first 30 projected years (a performance metrics function of class PM)
}
\examples{
loadABT()
AvC30(MSE_example)
}
