% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Performance_metrics.R
\name{AvgBr}
\alias{AvgBr}
\title{Average Br (spawning biomass relative to dynamic SSBMSY) over projection years 11-30 (a performance metrics function of class PM)}
\usage{
AvgBr(MSE, pp = 1)
}
\arguments{
\item{MSE}{An object of class MSE}
}
\value{
a matrix n Management procedures (MSE@nMP) rows and nsim (MSE@nsim) columns from \code{MSE}
}
\description{
Average Br (spawning biomass relative to dynamic SSBMSY) over projection years 11-30 (a performance metrics function of class PM)
}
\examples{
loadABT()
AvgBr(myMSE2)
}
