% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_documentation.R
\docType{data}
\name{Bad_Obs}
\alias{Bad_Obs}
\title{Generic bad quality observation error model}
\description{
Generic bad quality observation error model
}
\author{
Tom Carruthers \email{t.carruthers@oceans.ubc.ca}
}
\keyword{data}
