% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal.R
\name{betaconv}
\alias{betaconv}
\title{Calculate the beta distribution parameter beta from transformed space mean and standard deviations}
\usage{
betaconv(m, sd)
}
\arguments{
\item{m}{a fraction}

\item{sd}{a positive real number}
}
\value{
the beta distribution parameter alpha corresponding to \code{m, sd}
}
\description{
Calculate the beta distribution parameter beta from transformed space mean and standard deviations
}
\examples{
betaconv(0.25,0.1)
}
