% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Performance_metrics.R
\name{Br30}
\alias{Br30}
\title{Depletion (spawning biomass relative to dynamic BMSY) after projection yeare 30 (a performance metrics function of class PM)}
\usage{
Br30(MSE, pp = 1)
}
\arguments{
\item{MSE}{An object of class MSE}
}
\value{
a matrix with n Management procedures (MSE@nMP) and nsim (MSE@nsim) columns from \code{MSE}
}
\description{
Depletion (spawning biomass relative to dynamic BMSY) after projection yeare 30 (a performance metrics function of class PM)
}
\examples{
loadABT()
Br30(MSE_example)
}
