% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSE_functions.R
\name{Br30_Wt}
\alias{Br30_Wt}
\title{Get Br30 results accounting for OM weights}
\usage{
Br30_Wt(MSElist, q = 0.5)
}
\arguments{
\item{MSElist}{A list of the deterministic reference set MSE objects}

\item{q}{The percentile of the weighted distribution}
}
\description{
Get Br30 results accounting for OM weights
}
