% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Performance_metrics.R
\name{C10}
\alias{C10}
\title{Mean  catches over the first 10 projected years (a performance metrics function of class PM)}
\usage{
C10(MSE, pp = 1)
}
\arguments{
\item{MSE}{An object of class MSE}
}
\value{
a matrix with n Management procedures (MSE@nMP) and nsim (MSE@nsim) columns from \code{MSE}
}
\description{
Mean  catches over the first 10 projected years (a performance metrics function of class PM)
}
\examples{
loadABT()
C10(MSE_example)
}
