% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Performance_metrics.R
\name{C30}
\alias{C30}
\title{Mean catches over projected years 21-30 (a performance metrics function of class PM)}
\usage{
C30(MSE, pp = 1)
}
\arguments{
\item{MSE}{An object of class MSE}
}
\value{
a matrix with n Management procedures (MSE@nMP) and nsim (MSE@nsim) columns from \code{MSE}
}
\description{
Mean catches over projected years 21-30 (a performance metrics function of class PM)
}
\examples{
loadABT()
C30(MSE_example)
}
