% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\name{CDD_i2}
\alias{CDD_i2}
\title{A delay difference model fitted to INDEX 2 linked to Justin Cooke's harvest control rule (a management procedure of class MP).}
\usage{
CDD_i2(x, dset)
}
\arguments{
\item{x}{a simulation number.}

\item{dset}{a list of simulated data for use by management procedures.}
}
\value{
a TAC recommendation arising from \code{x, dset}.
}
\description{
A delay difference model fitted to INDEX 2 linked to Justin Cooke's harvest control rule (a management procedure of class MP).
}
\examples{
CDD_i2(1,dset_example_West)
sapply(1:10,CDD_i2,dset_example_West)
}
