% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OM_functions.R
\name{CheckOMI}
\alias{CheckOMI}
\title{Check data dimenstions of an OMI object}
\usage{
CheckOMI(OMI)
}
\arguments{
\item{OMI}{an operating model input object of class OMI.}
}
\value{
a printout / diagonstic of OMI data dimensions
}
\description{
Check data dimenstions of an OMI object
}
