% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\docType{data}
\name{Cooke_DD}
\alias{Cooke_DD}
\title{A delay difference model linked to Justin Cooke's harvest control rule (a management procedure of class ABT_MP).}
\format{An object of class \code{ABT_MP} of length 1.}
\usage{
Cooke_DD
}
\arguments{
\item{x}{a simulation number.}

\item{pset}{a list of simulated data for use by management procedures.}
}
\value{
a TAC recommendation arising from \code{x, pset}.
}
\description{
A delay difference model linked to Justin Cooke's harvest control rule (a management procedure of class ABT_MP).
}
\examples{
DD4010(1,pset_example)
sapply(1:10,DD4010,pset_example)
}
\keyword{datasets}

