% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSMPs.R
\name{CSMP}
\alias{CSMP}
\title{Combined Stock Management Procedure.}
\usage{
CSMP(
  x,
  dset,
  AS,
  Balpha = 1,
  Bbeta = 2,
  Falpha = 0.8,
  Fbeta = 0.5,
  i1CR = 1,
  i1SUV = 3,
  i2CR = 6,
  i2SUV = 7,
  theta = matrix(c(0.8, 0.3, 0.2, 0.7), nrow = 2)
)
}
\arguments{
\item{x}{a simulation number.}

\item{dset}{a list with two positions (1:East  2:West) of simulated data for use by management procedures.}

\item{AS}{the management area for which advice is being provided (1:East 2:West)}

\item{Balpha}{degree of TAC throttling with declining biomass}

\item{Bbeta}{non-linearity in TAC throttling with declining biomass}

\item{Falpha}{target fraction of F relative to FMSY}

\item{Fbeta}{degree of TAC throttling with increasing F}

\item{i1CR}{Catch rate index for stock 1 (East)}

\item{i1SUV}{Survey index for stock 1 (East)}

\item{i2CR}{Catch rate index for stock 2 (West)}

\item{i2SUV}{Survey index for stock 2 (West)}

\item{theta}{Assumed mixing matrix from Stock (row, E/W) to Area (column, E/W)}
}
\value{
a TAC recommendation arising from \code{x, dset}.
}
\description{
Combined Stock Management Procedure.
}
\examples{
ZeroC(1,dset_example_West)
sapply(1:10,ZeroC,dset_example_West)
}
