% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\name{CurC100}
\alias{CurC100}
\title{Current catches (a management procedure of class MP).}
\usage{
CurC100(x, dset)
}
\arguments{
\item{x}{a simulation number.}

\item{dset}{a list of simulated data for use by management procedures.}
}
\value{
a TAC recommendation arising from \code{x, dset}.
}
\description{
Current catches (a management procedure of class MP).
}
\examples{
CurC100(1,dset_example_West)
sapply(1:10,CurC100,dset_example_West)
}
