% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\name{CurC150}
\alias{CurC150}
\title{Current catches x 1.5 (a management procedure of class MP).}
\usage{
CurC150(x, dset)
}
\arguments{
\item{x}{a simulation number.}

\item{dset}{a list of simulated data for use by management procedures.}
}
\value{
a TAC recommendation arising from \code{x, dset}.
}
\description{
Current catches x 1.5 (a management procedure of class MP).
}
\examples{
CurC150(1,dset_example_West)
sapply(1:10,CurC150,dset_example_West)
}
