% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\name{CurC25}
\alias{CurC25}
\title{One quarter current catches (a management procedure of class MP).}
\usage{
CurC25(x, dset)
}
\arguments{
\item{x}{a simulation number.}

\item{dset}{a list of simulated data for use by management procedures.}
}
\value{
a TAC recommendation arising from \code{x, dset}.
}
\description{
One quarter current catches (a management procedure of class MP).
}
\examples{
CurC25(1,dset_example_West)
sapply(1:10,CurC25,dset_example_West)
}
