% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\name{CurC50}
\alias{CurC50}
\title{Half current catches (a management procedure of class MP).}
\usage{
CurC50(x, dset)
}
\arguments{
\item{x}{a simulation number.}

\item{dset}{a list of simulated data for use by management procedures.}
}
\value{
a TAC recommendation arising from \code{x, dset}.
}
\description{
Half current catches (a management procedure of class MP).
}
\examples{
CurC50(1,dset_example_West)
sapply(1:10,CurC50,dset_example_West)
}
