% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\name{CurC75}
\alias{CurC75}
\title{Three quarters current catches (a management procedure of class MP).}
\usage{
CurC75(x, dset)
}
\arguments{
\item{x}{a simulation number.}

\item{dset}{a list of simulated data for use by management procedures.}
}
\value{
a TAC recommendation arising from \code{x, dset}.
}
\description{
Three quarters current catches (a management procedure of class MP).
}
\examples{
CurC75(1,dset_example_West)
sapply(1:10,CurC75,dset_example_West)
}
