% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal.R
\name{cv}
\alias{cv}
\title{Calculate coefficient of variance}
\usage{
cv(x)
}
\arguments{
\item{x}{a vector of length 2 or more}
}
\value{
the coefficient of variation of \code{arr}
}
\description{
Calculate coefficient of variance
}
\examples{
cv(runif(10))
}

