% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\docType{data}
\name{DD}
\alias{DD}
\title{A rapid 2 parameter observation error only delay difference model conditioned on effort and parameterized with UMSY and MSY leading (a management procedure of class ABT_MP).}
\format{An object of class \code{ABT_MP} of length 1.}
\usage{
DD
}
\arguments{
\item{x}{a simulation number.}

\item{pset}{a list of simulated data for use by management procedures.}
}
\value{
a TAC recommendation arising from \code{x, pset}.
}
\description{
A rapid 2 parameter observation error only delay difference model conditioned on effort and parameterized with UMSY and MSY leading (a management procedure of class ABT_MP).
}
\examples{
DD(1,pset_example)
sapply(1:10,DD,pset_example)
}
\keyword{datasets}

