% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\name{DD2}
\alias{DD2}
\title{A rapid 2 parameter observation error only delay difference model conditioned on effort and parameterized with UMSY and MSY leading similar to DD but removes all NA values from data history (a management procedure of class ABT_MP).}
\usage{
DD2(x, pset)
}
\arguments{
\item{x}{a simulation number.}

\item{pset}{a list of simulated data for use by management procedures.}
}
\value{
a TAC recommendation arising from \code{x, pset}.
}
\description{
A rapid 2 parameter observation error only delay difference model conditioned on effort and parameterized with UMSY and MSY leading similar to DD but removes all NA values from data history (a management procedure of class ABT_MP).
}
\examples{
DD2(1,pset_example)
sapply(1:10,DD2,pset_example)
}

