% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\name{DD_i3_4010}
\alias{DD_i3_4010}
\title{A rapid 3 parameter observation error only delay difference model linked to 40-10 harvest control rule FOR INDEX 3 (GOM_LAR_SUV) conditioned on effort and parameterized with UMSY and MSY leading (a management procedure of class MP).}
\usage{
DD_i3_4010(x, dset, checkfit = F)
}
\arguments{
\item{x}{a simulation number.}

\item{dset}{a list of simulated data for use by management procedures.}

\item{checkfit}{logical: should fitting diagnostics be plotted}
}
\value{
a TAC recommendation arising from \code{x, dset}.
}
\description{
A rapid 3 parameter observation error only delay difference model linked to 40-10 harvest control rule FOR INDEX 3 (GOM_LAR_SUV) conditioned on effort and parameterized with UMSY and MSY leading (a management procedure of class MP).
}
\examples{
DD_i3_4010(1,dset_example_West)
sapply(1:10,DD_i3_4010,dset_example_East)
}
