% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OM_functions.R
\name{Deterministic}
\alias{Deterministic}
\title{Remove process error from an operating model projections (no recruitment deviations)}
\usage{
Deterministic(OM)
}
\arguments{
\item{OM}{an operating model object of class OM.}
}
\value{
an object of class OM without process errors \code{x, dset}.
}
\description{
Remove process error from an operating model projections (no recruitment deviations)
}
\examples{
OM_1d<-Deterministic(OM_1)
}
