% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Performance_metrics.R
\name{DNC}
\alias{DNC}
\title{Relative SSB (SSB relative to zero fishing) in final projection year (a performance metrics function of class PM)}
\usage{
DNC(MSE, pp = 1)
}
\arguments{
\item{MSE}{An object of class MSE}
}
\value{
a matrix with n Management procedures (MSE@nMP) and nsim (MSE@nsim) columns from \code{MSE}
}
\description{
Relative SSB (SSB relative to zero fishing) in final projection year (a performance metrics function of class PM)
}
\examples{
loadABT()
DNC(MSE_example)
}
