% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_documentation.R
\docType{data}
\name{dset}
\alias{dset}
\title{A simulated data set}
\description{
\describe{
\item{Cobs}{A matrix of [sim x nyear]}
\item{Iobs}{A 3D array of observed relative abundance indices dimensions [sim x index x year]}
\item{K}{A vector nsim long of von Bertalanffy growth parameter K estimates}
\item{Linf}{A vector nsim long of growth parameter Linf estimates (maximum length)}
\item{t0}{A vector nsim long of growth parameter t0 estimates}
\item{M}{A matrix of natural mortality rate at age [sim x age]}
\item{Bt}{A vector nsim long of current biomass estimates}
\item{MSY}{A vector nsim long of MSY estimates}
\item{BMSY}{A vector nsim long of BMSY estimates}
\item{UMSY}{A vector nsim long of UMSY estimates}
\item{a}{A vector nsim long of Length-weight conversion parameter a W=aL^b}
\item{b}{A vector nsim long of Length-weight conversion parameter b W=aL^b}
\item{nages}{An integer value specifying the total number of age classes}
\item{ageM}{A vector nsim long of estimates of age at 50 per cent maturity}
\item{M}{A matrix of maturity at age [sim x age]}
\item{Bt_PI}{A vector nsim long of current biomass estimates known perfectly (PI: Perfect Information)}
\item{UMSY_PI}{A vector nsim long of UMSY nown perfectly (PI: Perfect Information)}
\item{CAA}{A 3D array of catch at age observations [sim x age x year]}
\item{CAL}{A vector of lengths representing the start-end points of the length bins corresponding to CAL data]}
\item{MPrec}{A vector nsim long of the previous recommendations of the management procedure}
}
}
\author{
Tom Carruthers \email{t.carruthers@oceans.ubc.ca}
}
\keyword{data}
