% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\name{EMP1e}
\alias{EMP1e}
\title{Example Management Procedure 1 east using the JP_LL_NE}
\usage{
EMP1e(x, dset, Jtarg = 4.8, TACadj = 0.1, thresh = 0.4)
}
\arguments{
\item{x}{a simulation number.}

\item{dset}{a list of simulated data for use by management procedures.}
}
\value{
a TAC recommendation arising from \code{x, dset}.
}
\description{
Example Management Procedure 1 east using the JP_LL_NE
}
\examples{
EMP1e(1,dset_example_West)
sapply(1:10,EMP1e,dset_example_West)
}
