% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\name{EMP1w}
\alias{EMP1w}
\title{Example Management Procedure 1 west using the GOM_LAR_SUV}
\usage{
EMP1w(x, dset, Jtarg = 0.66, TACadj = 0.1, thresh = 0.4)
}
\arguments{
\item{x}{a simulation number.}

\item{dset}{a list of simulated data for use by management procedures.}
}
\value{
a TAC recommendation arising from \code{x, dset}.
}
\description{
Example Management Procedure 1 west using the GOM_LAR_SUV
}
\examples{
EMP1w(1,dset_example_West)
sapply(1:10,EMP1w,dset_example_West)
}
