% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\name{EMP2e}
\alias{EMP2e}
\title{Example Management Procedure 2 east.}
\usage{
EMP2e(x, dset, Jtarg = 4.8, lup = 0.05, ldown = 0.15, pup = 0.05, pdown = 0.15)
}
\arguments{
\item{x}{a simulation number.}

\item{dset}{a list of simulated data for use by management procedures.}
}
\value{
a TAC recommendation arising from \code{x, dset}.
}
\description{
Example Management Procedure 2 east.
}
\examples{
EMP2e(1,dset_example_West)
sapply(1:10,EMP2e,dset_example_West)
}
