% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\name{EMP2w}
\alias{EMP2w}
\title{Example Management Procedure 2 west.}
\usage{
EMP2w(x, dset, Jtarg = 0.66, lup = 0.05, ldown = 0.15, pup = 0.05,
  pdown = 0.15)
}
\arguments{
\item{x}{a simulation number.}

\item{dset}{a list of simulated data for use by management procedures.}
}
\value{
a TAC recommendation arising from \code{x, dset}.
}
\description{
Example Management Procedure 2 west.
}
\examples{
EMP2w(1,dset_example_West)
sapply(1:10,EMP2w,dset_example_West)
}
