% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\docType{data}
\name{Fadapt}
\alias{Fadapt}
\title{An adaptive surplus production management procedure that attempts to identify the slope of changes in surplus production and biomass to locate a gradient of zero (a management procedure of class ABT_MP).}
\format{An object of class \code{ABT_MP} of length 1.}
\usage{
Fadapt
}
\arguments{
\item{x}{a simulation number.}

\item{pset}{a list of simulated data for use by management procedures.}

\item{yrsmth}{the number of years over which to quantify slope in surplus production with biomass}

\item{gg}{the responsiveness of the MP with respect to the slope}

\item{FMSY_M}{the assumed ratio of FMSY to M}
}
\value{
a TAC recommendation arising from \code{x, pset, yrsmth, alp, bet}.
}
\description{
An adaptive surplus production management procedure that attempts to identify the slope of changes in surplus production and biomass to locate a gradient of zero (a management procedure of class ABT_MP).
}
\examples{
Fadapt(1,pset_example)
sapply(1:10,Fadapt,pset_example)
}
\keyword{datasets}

