% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\name{Fadapt_i7}
\alias{Fadapt_i7}
\title{An adaptive surplus production management procedure using INDEX 7 (GOM_LAR_SUV) that attempts to identify the slope of changes in surplus production and biomass to locate a gradient of zero (a management procedure of class MP).}
\usage{
Fadapt_i7(x, dset)
}
\arguments{
\item{x}{a simulation number.}

\item{dset}{a list of simulated data for use by management procedures.}

\item{yrsmth}{the number of years over which to quantify slope in surplus production with biomass}

\item{gg}{the responsiveness of the MP with respect to the slope}

\item{FMSY_M}{the assumed ratio of FMSY to M}
}
\value{
a TAC recommendation arising from \code{x, dset}.
}
\description{
An adaptive surplus production management procedure using INDEX 7 (GOM_LAR_SUV) that attempts to identify the slope of changes in surplus production and biomass to locate a gradient of zero (a management procedure of class MP).
}
\examples{
Fadapt_i7(1,dset_example_East)
sapply(1:10,Fadapt_i7,dset_example_East)
}
