% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_documentation.R
\docType{data}
\name{Good_Obs}
\alias{Good_Obs}
\title{Generic good quality observation error model}
\description{
Generic good quality observation error model
}
\author{
Tom Carruthers \email{t.carruthers@oceans.ubc.ca}
}
\keyword{data}

