% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCRs.R
\name{HCR60_20}
\alias{HCR60_20}
\title{A 60-20 harvest control rule based on B/BMSY}
\usage{
HCR60_20(Brel)
}
\arguments{
\item{Brel}{Improper fraction: an estimate of Biomass relative to BMSY}
}
\value{
a TAC or TAE adjustement factor.
}
\description{
A 60-20 harvest control rule based on B/BMSY
}
\examples{
Brel<-seq(0,1.5,length.out=200)
plot(Brel,HCR60_20(Brel),xlab="B/BMSY",ylab="HCR adjustement factor",main="A 60-20 harvest control rule",type='l',col='blue')
abline(v=c(0.2,0.6),col='red',lty=2)
}
\references{
Pretty generic fishery control rule
}
\author{
T. Carruthers
}
