% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCRs.R
\name{HCRlin}
\alias{HCRlin}
\title{Generic linear harvest control rule based on B/BMSY}
\usage{
HCRlin(Brel, LRP, TRP)
}
\arguments{
\item{Brel}{Improper fraction: an estimate of Biomass relative to BMSY}

\item{LRP}{Improper fraction: the Limit Reference Point, a biomass level below which the adjustment is zero (no fishing)}

\item{TRP}{Improper fraction: the Target Reference Point, a biomass level above which the adjustment is 1 (no adjustment)}
}
\value{
a TAC or TAE adjustement factor.
}
\description{
Generic linear harvest control rule based on B/BMSY
}
\examples{
Brel<-seq(0,1.5,length.out=200)
plot(Brel,HCRlin(Brel,0.1,0.4),xlab="B/BMSY",ylab="HCR adjustement factor",main="Basically a 40-10 harvest control rule",type='l',col='blue')
abline(v=c(0.1,0.4),col='red',lty=2)
}
\references{
Pretty generic fishery control rule
}
\author{
T. Carruthers
}
