% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M3_tools.R
\name{IndLHDs}
\alias{IndLHDs}
\title{Make negative log-likelihood components table}
\usage{
IndLHDs(outs, OMIs, OMnams)
}
\arguments{
\item{outs}{a list of M3 output files created from M3read()}

\item{OMIs}{a list of M3 OMI objects}

\item{OMnams}{a character string of operating model names}
}
\value{
a list of tables of negative log likelihood components for CPUE and FI indices
}
\description{
Make negative log-likelihood components table
}
