% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\docType{data}
\name{Islope1}
\alias{Islope1}
\title{An index-based management procedure that incrementally changes TAC to achieve a stable index(a management procedure of class ABT_MP).}
\format{An object of class \code{ABT_MP} of length 1.}
\usage{
Islope1
}
\arguments{
\item{x}{a simulation number.}

\item{pset}{a list of simulated data for use by management procedures.}

\item{yrsmth}{the number of years over which to compare index observations}

\item{lambda}{a scalar which determines the responsiveness of the adjustment}

\item{xx}{alters target catch as a fraction of historical catch TACstar = (1-xx)muC}
}
\value{
a TAC recommendation arising from \code{x, pset, yrsmth, lamba, xx}.
}
\description{
An index-based management procedure that incrementally changes TAC to achieve a stable index(a management procedure of class ABT_MP).
}
\examples{
Islope1(1,pset_example)
sapply(1:10,Islope1,pset_example)
}
\keyword{datasets}

