% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\docType{data}
\name{Islope1_i4}
\alias{Islope1_i4}
\title{An index-based (INDEX 4) management procedure that incrementally changes TAC to achieve a stable index(a management procedure of class MP).}
\format{An object of class \code{MP} of length 1.}
\usage{
Islope1_i4
}
\arguments{
\item{x}{a simulation number.}

\item{dset}{a list of simulated data for use by management procedures.}

\item{yrsmth}{the number of years over which to compare index observations}

\item{xx}{alters target catch as a fraction of historical catch TACstar = (1-xx)muC}

\item{lambda}{a scalar which determines the responsiveness of the adjustment}
}
\value{
a TAC recommendation arising from \code{x, dset}.
}
\description{
An index-based (INDEX 4) management procedure that incrementally changes TAC to achieve a stable index(a management procedure of class MP).
}
\examples{
Islope1_i4(1,dset_example_East)
sapply(1:10,Islope1_i4,dset_example_East)
}
\keyword{datasets}

