% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSE_functions.R
\name{Join_Results}
\alias{Join_Results}
\title{Compile the performance results for a complete set of MSE runs (all reference and robustness operating models)}
\usage{
Join_Results(ResList)
}
\arguments{
\item{ResList}{A list of results objects compiled by the function Results_compiler()}
}
\description{
Compile the performance results for a complete set of MSE runs (all reference and robustness operating models)
}
