% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M3_tools.R
\name{LHtabs}
\alias{LHtabs}
\title{Make negative log-likelihood components table}
\usage{
LHtabs(outs, OMIs, OMnos, OMnams, rnd = 0)
}
\arguments{
\item{outs}{list of output files each created from M3read()}

\item{OMIs}{list of OMI file paired to each position in outs()}

\item{OMnos}{vector of OM numbers}

\item{OMnams}{vector of OM names}

\item{rnd}{integer numbers in tables are rounded to rnd decimal places}
}
\value{
a list 2 elements long of (1) weighted and (2) unweighted negative log-likelihoods
}
\description{
Make negative log-likelihood components table
}
