% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M3_tools.R
\name{M3read}
\alias{M3read}
\title{Reads the outputs of an M3 model fit into a list}
\usage{
M3read(OMDir = "C:/M3", quiet = T)
}
\arguments{
\item{OMdir}{a character string representing the location of an M3 folder containing M3.rep, M3.par and M3.cor files}
}
\value{
runs a list object of M3 outputs from folder \code{OMdir}
}
\description{
Reads the outputs of an M3 model fit into a list
}
\examples{
M3read("C:/ABT-MSE/M3")
}
