% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M3_tools.R
\name{M3write}
\alias{M3write}
\title{Create an M3.dat input file from an object of class OMI}
\usage{
M3write(OMI, OMdir = "C:/M3")
}
\arguments{
\item{OMI}{and object of class OMI}

\item{OMdir}{a character string representing the location of an M3 folder containing M3.exe and stats.cxx files}
}
\value{
creates an M3.dat file of the correct format in \code{OMdir}
}
\description{
Create an M3.dat input file from an object of class OMI
}
\examples{
M3write(OMI_example,C:/ABT-MSE/M3)
}

