% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M3_tools.R
\name{make_fit_reports}
\alias{make_fit_reports}
\title{Make M3 fitting reports in all directors of a folder}
\usage{
make_fit_reports(dirs = "C:/M3", addlab = FALSE)
}
\arguments{
\item{dirs}{character: an folder containing subfolders that are fitted M3 OMs}

\item{addlab}{logical: whether OMI objects should be looked for in the directory to include more detailed OM info}
}
\value{
a set of pdf reports in each subdirectory
}
\description{
Make M3 fitting reports in all directors of a folder
}
