% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\name{MCD_i7}
\alias{MCD_i7}
\title{A simple MP that uses an index (INDEX 7 - JPN_LL_EAtl_Med) to derive depletion then fishing at average catch x depletion x 2(a management procedure of class MP).}
\usage{
MCD_i7(x, dset)
}
\arguments{
\item{x}{a simulation number.}

\item{dset}{a list of simulated data for use by management procedures.}
}
\value{
a TAC recommendation arising from \code{x, dset}.
}
\description{
A simple MP that uses an index (INDEX 7 - JPN_LL_EAtl_Med) to derive depletion then fishing at average catch x depletion x 2(a management procedure of class MP).
}
\examples{
MCD_i7(1,dset_example_East)
sapply(1:10,MCD_i7,dset_example_East)
}
