% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal.R
\name{mconv}
\alias{mconv}
\title{Calculate the lognormal mean from logspace mean and standard deviations}
\usage{
mconv(m, sd)
}
\arguments{
\item{m}{a positive real number}

\item{sd}{a positive real number}
}
\value{
the lognormal mean corresponding to \code{m, sd}
}
\description{
Calculate the lognormal mean from logspace mean and standard deviations
}
\examples{
mconv(0.5,0.2)
}

