% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\docType{data}
\name{MeanC}
\alias{MeanC}
\title{A simple average catch MP.}
\format{An object of class \code{MP} of length 1.}
\usage{
MeanC
}
\arguments{
\item{x}{a simulation number.}

\item{dset}{a list of simulated data for use by management procedures.}
}
\value{
a TAC recommendation arising from \code{x, dset}.
}
\description{
A simple average catch MP.
}
\examples{
MeanC(1,dset_example_East)
}
\keyword{datasets}

