% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSE_Object.R
\docType{class}
\name{MSE-class}
\alias{MSE-class}
\title{An S4 class object that contains all MSE outputs}
\description{
\describe{
\item{Name}{The name of the MSE}
\item{Date}{The date and time the object was created}
\item{Author}{Who made the object}
\item{Notes}{Anything noteworthy}
\item{PrimarySource}{The primary origins of these data}
\item{nsim}{The number of simulations}
\item{npop}{number of stocks (e.g. 2, East-West)}
\item{nages}{number of ages (e.g. 18: 1-17 and 18+)}
\item{nyears}{number of years of simulation by statistical catch at length}
\item{nHyears}{number of historical years for stock reduction analysis prior ot statistical catch at length}
\item{nsubyears}{number of subyears (e.g. 4 quarters)}
\item{nareas}{number of areas}
\item{proyears}{number of projection years for closed-loop MSE analysis}
\item{nlen}{number of length classes}
\item{lenbins}{vector of length bins that define length classes}
\item{interval}{how often (number of years) assessments/MPs are implemented (e.g. interval=3 means 2015 assessment, 2016-2017 no assessment, 2018 assessment etc)}
\item{nma}{number of movement age classes}
\item{ma}{an array of movement age classes [npop x nage]}
\item{mulen}{a vector of mean length in each length bin [nage]}
\item{Cimp}{a vector of imprecision in catch observations (lognormal sd) by simulation [nsim]}
\item{Cb}{a vector representing the bias in catch observations}
\item{Cerr}{a matrix of the catch mulitpliers applied to true simulated catch the product of Cimp and Cb [nsim x nyears]}
\item{Iimp}{currently unused}
\item{Ib}{currently unused}
\item{Ierr}{currently unused}
\item{nCAAobs}{a vector of the number of annual catch-at-age observations by simulation [nsim]}
\item{nCALobs}{a vector of the number of annual catch-at-length observations by simulation [nsim]}
\item{Lcv}{currently unused}
\item{Mb}{a vector of biases in natural mortality rate by simulation [nsim]}
\item{Kb}{a vector of biases in von-Bert growth parameter K by simulation [nsim]}
\item{t0b}{a vector of biases in von-Bert growth parameter t0 by simulation [nsim]}
\item{Mb}{a vector of biases in von-Bert maximum length parameter by simulation [nsim]}
\item{LFCb}{currently unused}
\item{LFSb}{currently unused}
\item{FMSYb}{a vector of biases in fishing mortality rate at MSY by simulation [nsim]}
\item{FMSY_Mb}{a vector of biases in the ratio of fishing mortality rate at MSY to natural mortality rate by simulation [nsim]}
\item{BMSY_B0b}{currently unused}
\item{ageMb}{a vector of biases in the age at 50 per cent maturity [nsim]}
\item{Dimp}{currently unused}
\item{Db}{currently unused}
\item{Derr}{currently unused}
\item{Btimp}{a vector of imprecision in current biomass observations (lognormal sd) by simulation [nsim]}
\item{Btb}{a vector representing the bias in current biomass observations}
\item{Bterr}{a matrix of the mulitpliers applied to true simulated biomass the product of Btimp and Btb [nsim x nyears]}
\item{Ftimp}{currently unused}
\item{Ftb}{currently unused}
\item{Fterr}{currently unused}
\item{hb}{currently unused}
\item{Reccbcv}{currently unused}
\item{IMSYb}{a vector of biases in the index at MSY by simulation [nsim]}
\item{IMSYb}{a vector of biases in MSY (a reference catch level) by simulation [nsim]}
\item{IMSYb}{a vector of biases in BMSY (a reference biomass level) by simulation [nsim]}
\item{C}{a 4D array containing true simulated catches [MP x nsim x nstocks x nyears]}
\item{D}{a 4D array containing true simulated stock depletion (SSB/SSB0) [MP x nsim x nstocks x nyears]}
\item{B_BMSY}{a 4D array containing true simulated biomass relative to BMSY [MP x nsim x nstocks x nyears]}
\item{F_FMSY}{a 4D array containing true simulated fishing mortality rate relative to FMSY [MP x nsim x nstocks x nyears]}
\item{B}{currently unused}
\item{SSB}{a 4D array containing true simulated spawning biomass [MP x nsim x nstocks x nyears]}
\item{SSB0}{a 2D array containing unfished spawnign biomass by simulation and stock [nsim x nstocks]}
\item{SSB0proj}{a 3D array containing unfished spawning biomass by simulation, stock and projection year [nsim x nstocks x proyears]}
\item{TAC}{a 4D array containing the TAC recommendations [nsim x MP x stock x proyear]}
\item{nMPs}{an integer number representing the number of MPs in the MSE}
\item{Snames}{a character vector naming the stocks [nstocks]}
\item{area_defs}{a list of area definitions for graphing (lons and lats describing the polygon)}
\item{areanams}{a character vector of area names}
\item{Istats}{an array of index fit statistics nsim x nind x 4 (beta, AC, sd, cor)}
\item{MPs}{a list object containing the names of the MPs [nMPs]}
}
}
