% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSE_auto_report.R
\name{MSE_report}
\alias{MSE_report}
\title{Make standardized MSE MP performance report}
\usage{
MSE_report(MSE, dir = NA, Author = "Anon", introtext = NA,
  filenam = "MSE_report")
}
\arguments{
\item{MSE}{An object of class MSE return from a function call MSE<-new('MSE',...)}

\item{dir}{A character string that is the directory where the report will be saved (default is current working directory)}

\item{Author}{A character string; the author of the report}

\item{introtext}{A character string; optional; Any relevant text the author would like to include about the MSE}

\item{filenam}{A character string: optional: a specific name of the report file so for Daves_MP_perf.html, filenam="Daves_MP_perf"}
}
\value{
an html report in the directory dir
}
\description{
Make standardized MSE MP performance report
}
