% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSE_functions.R
\name{multiMSE}
\alias{multiMSE}
\title{Run whole MSE runs in parallel}
\usage{
multiMSE(
  OMdirs,
  MPs = MPs,
  nsim = 12,
  Obs,
  rebuildOMs = T,
  outfolder,
  seed = 1,
  deterministic = F,
  Recs = Recs,
  Design = Design
)
}
\arguments{
\item{MPs}{A vector of management procedures for analysis}

\item{nsim}{The number of simulations}

\item{Obs}{an observation error model of class 'Obs' e.g. Perfect_Obs}

\item{rebuildOMs}{Logical: should the OMs be rebuilt or should those in the directory be used directly}

\item{outfolder}{the folder where output MSE objects should be placed}

\item{seed}{the random seed for the MSE}

\item{deterministic}{Logical: should the deterministic OMd be used}

\item{OMdir}{A vector of operating model directories (e.g. those found in abft-mse/Objects/OMs/)}
}
\description{
Run whole MSE runs in parallel
}
