% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_documentation.R
\docType{data}
\name{Obs_example}
\alias{Obs_example}
\title{#' An example observation error model (for everthing but indices)}
\description{
#' An example observation error model (for everthing but indices)
}
\author{
Tom Carruthers \email{t.carruthers@oceans.ubc.ca}
}
\keyword{data}

