% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Performance_metrics.R
\name{OFT}
\alias{OFT}
\title{Over-fished Trend, the slope in SSB at the origin Good trend: 1 minus probability of negative trend (Br31 – Br35) and Br30 is less than 1 (a performance metrics function of class PM)}
\usage{
OFT(MSE, pp = 1)
}
\arguments{
\item{MSE}{An object of class MSE}
}
\value{
a matrix of 1's and zeros, n Management procedures (MSE@nMP) rows and nsim (MSE@nsim) columns from \code{MSE}
}
\description{
Over-fished Trend, the slope in SSB at the origin Good trend: 1 minus probability of negative trend (Br31 – Br35) and Br30 is less than 1 (a performance metrics function of class PM)
}
\examples{
loadABT()
OFT(myMSE2)
}
