% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OM_object.R
\docType{class}
\name{OM-class}
\alias{OM-class}
\title{An S4 class object that contains all Operating Model estimates in the format for M3 operating model estimation}
\description{
\describe{
\item{Name}{The name of the operating model input}
\item{Date}{The date and time the object was created}
\item{Author}{Who made the object}
\item{Notes}{Anything noteworthy}
\item{PrimarySource}{The primary origins of these data}
\item{nsim}{The number of simulations}
\item{npop}{number of stocks (e.g. 2, East-West)}
\item{nages}{number of ages (e.g. 18: 1-17 and 18+)}
\item{nyears}{number of years of simulation by statistical catch at length}
\item{nHyears}{number of historical years for stock reduction analysis prior ot statistical catch at length}
\item{nsubyears}{number of subyears (e.g. 4 quarters)}
\item{nareas}{number of areas}
\item{proyears}{number of projection years for closed-loop MSE analysis}
\item{nlen}{number of length classes}
\item{lenbins}{vector of length bins that define length classes}
\item{interval}{how often (number of years) assessments/MPs are implemented (e.g. interval=3 means 2015 assessment, 2016-2017 no assessment, 2018 assessment etc)}
\item{nma}{number of movement age classes}
\item{ma}{an array of movement age classes [npop x nage]}
\item{Rectype}{a matrix specifying the recruitment type (a stock can have multiple types of recruitment over the projected time period) by stock [type, stock]}
\item{h}{a vector nstocks long specifying the steepness of the B-H stock recruitment curve (currently unused)}
\item{recgrad}{a matrix specifying a possible future gradient in recruitment expressed as a per cent per year [stock x 2 (upper / lower)]}
\item{Recind}{an array specifyign which historical recruitment type to use in each projected year [nsim x nstock x projection year]}
\item{Recpars}{an array of stock recruitment parameters by recruitment type [nsim x nstock x rectype (Hockey stock / B-H) x 2 (h or jointpoint/R0)]}
\item{Reccv}{a vector nstock long of recruitment variability by stock}
\item{AC}{a vector nstock long of recruitment autocorrelation}
\item{Recsubyr}{a vector nstock long specifying which subyear recruitment occurs in}
\item{Linfmu}{a vector nstock long specifying the mean maximum length of individuals}
\item{Kmu}{a vector nstock long specifying the mean growth parameter K}
\item{t0}{a vector nstock long specifying the theoretical age at length zero}
\item{a}{a vector nstock long, the weight-length parameter a W=aL^b}
\item{b}{a vector nstock long, the weight-length parameter b W=aL^b}
\item{Len_age}{an array of length at age by stock [stock x age]}
\item{Wt_age}{an array of weight at age by stock [stock x age]}
\item{ageM}{an array of maturity at age by stock [stock x age]}
\item{D}{current stock depletion by stock (currently unused)}
\item{Size_area}{a vector nareas long, the relative size of each area}
\item{mov}{a very large array storing simulated movement by stock [sim x stock x age x subyear x movtype x from area x to area]}
\item{movIndex}{a vector of movement types [nyears+proyears]}
\item{nfleets}{the number of fishing fleets}
\item{iALK}{an array representing the inverse age-length key (conditional probability of length given age) [sim x stock x year x age x length class]}
\item{hZ}{an array of historical mortality rate estimates [stock, year, subyear, age, area]}
\item{HCobs}{an array of historical catch observations [year, subyear, age, area]}
\item{Cobs}{an array of catch observations [year, subyear, area, fleet]}
\item{E}{an array of standardized effort (partial effort)[sim, fleet, year, subyear, area]}
\item{qE}{a vector nfleets long, the catchability of the fleets}
\item{qI}{a vector nI long of catchabilities by fishery independent index}
\item{qCPUE}{a vector nCPUE long of catchabilities by fishery CPUE index}
\item{sel}{an array of selectivities [sim, fleets, length class]}
\item{selpars}{array of selectivity parameters (currently unused)}
\item{Fmod}{array of season x area mean F parameters}
\item{FDY}{array of annual F deviations [sim, year, quarter, area]}
\item{mat}{array of maturity [sim, stock, age, year]}
\item{M}{array of natural mortality rate [sim, stock, age, year]}
\item{Recdevs}{of historical recruitment deviations [sim, stock, year]}
\item{R0}{matrix of unfished recruitment [sim, stock, year]}
\item{Recpar}{matrix of recruitment pars [sim, stock, year]}
\item{Rectypes}{matrix of recruitment types [stock, year]}
\item{MSY}{a matrix of MSY values per simulation and stock [sim x stock]}
\item{BMSY}{a matrix of Biomass at MSY values per simulation and stock [sim x stock]}
\item{VBMSY}{a matrix of vulnerable Biomass at MSY values per simulation and stock [sim x stock]}
\item{SSBMSY}{a matrix of spawning biomass at MSY values per simulation and stock [sim x stock]}
\item{FMSY}{a matrix of fishing exploitation rate at MSY values per simulation and stock [sim x stock]}
\item{UMSY}{a matrix of fishing harvest rate at MSY values per simulation and stock [sim x stock]}
\item{FMSYa}{a matrix of apical (coresponding with age that is most vulnerable) fishing exploitation rate at MSY values per simulation and stock [sim x stock]}
\item{SSBMSY_SSB0}{a matrix of spawning biomass at MSY relative to unfished, values per simulation and stock [sim x stock]}
\item{nydist}{the number of years used to determine equilibrium spatial distribution}
\item{Snames}{a vector nstock long naming each stock}
\item{area_defs}{a list of area definitions for graphing (lons and lats describing the polygon)}
\item{areanams}{a character vector of area names}
\item{Ibeta_ignore}{logical, should hyperstability of indices be included?}
\item{qinc}{real number, percentage annual changes in catchability affecting the CPUE indices}
\item{SSBpR}{matrix simulation by stock}
\item{seed}{a random seed for generation of simulations to ensure reproducibility}
}
}
