% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Objects.R
\docType{class}
\name{OMI-class}
\alias{OMI-class}
\title{An S4 class object that contains all Operating Model Inputs in the format for M3 operating model estimation}
\description{
\describe{
\item{Name}{The name of the operating model input}
\item{Date}{The date and time the object was created}
\item{Author}{Who made the object}
\item{Notes}{Anything noteworthy}
\item{PrimarySource}{The primary origins of these data}
\item{OMfactors}{The various levels and factors of this operating model}
\item{years}{The real year range of the operating model}
\item{Hyears}{The historical year range of the operating model}
\item{areas}{The codes of the various geographical areas of the model}
\item{areanams}{Full names of the various geographical areas}
\item{area_defs}{A list of the polygons representing each of the areas (for plotting)}
\item{Fleets}{A list object containing the definitions of the fleets}
\item{A2}{Richards growth curve parameter}
\item{L1}{Richards growth curve parameter}
\item{L2}{Richards growth curve parameter}
\item{K}{Richards growth curve parameter}
\item{p}{Richards growth curve parameter}
\item{t0}{VB growth curve parameter}
\item{Linf}{VB growth curve parameter}
\item{Lvar_a}{Length variability at length parameter}
\item{Lvar_b}{Length variability at length parameter}
\item{ageM}{Age at 50 percent maturity (not currently specified)}
\item{ageMsd}{slope in the maturity ogive (not currently specified)}
\item{mat}{maturity at age ogive [stock, age class, year]}
\item{lenbins}{vector of length bins that define length classes}
\item{nHy}{number of historical years (the period of stock reduction analysis)}
\item{ny}{number of years for the statistical catch at length analysis}
\item{ns}{number of subyears (e.g. 4 quarters)}
\item{np}{number of stocks (e.g. 2, East-West)}
\item{na}{number of ages (e.g. 18: 1-17 and 18+)}
\item{nr}{number of areas}
\item{nf}{number of fleets}
\item{nl}{number of length classes}
\item{nRPT}{number of recapture time periods (currently unused)}
\item{RPTind}{the correct release subyear for a recapture in a given subyear at a particular recapture period (currently unused)}
\item{sdur}{the relative size of each subyear duration e.g. c(0.25,0.25,0.25,0.25) for a quarterly model}
\item{nydist}{number of years used to initialize the spatial distribution of the movement age classes}
\item{mulen}{mean length of individuals in each length class}
\item{RDblock}{a vector nyears long indicating which estimated recruitment to use in each year}
\item{nRD}{the number of estimated recruitments allocated to year by RDblock}
\item{iALK}{inverse age-length key, the conditional probability of a length class given an age class [stock, year, age class, length class]}
\item{lwa}{length-weight parameter a W=aL^b(currently unused)}
\item{lwb}{length-weight parameter b W=aL^b(currently unused)}
\item{len_age}{an array of length-at-age [stock, age, year]}
\item{wt_age}{an array of weight-at-age [stock, age, year]}
\item{wt_len}{a matrix of weight-at-length [stock, len]}
\item{Fec_len}{a matrix of Fecundity-at-length [stock, len]}
\item{Fec}{an array of spawning biomass-at-age [stock, age, year]}
\item{SRtype}{a character vector denoting the form of the stock-recruitment relationship: BH Beverton Holt, HS Hockey Stick [stock]}
\item{steep}{a vector of steepness by stock [stock]}
\item{SSBpR}{a vector of unfished spawning stock biomass per recruit [stock]}
\item{spawns}{the spawning season for each stock (a vector nstocks long)}
\item{canspawn}{an array [nstock, nareas] indicating which areas each stock can spawn in (then calculated by the fraction of SSB in each area during spawns)}
\item{Ma}{mortality at age [nstock, nages]}
\item{nCobs}{the number of catch observations in the statistical catch at length phase}
\item{Cobs}{a dataframe of catch observations, one row per observation [year, subyear, area, fleet observation]}
\item{nCPUEq}{the number of CPUE indices used in the model fitting}
\item{nCPUEobs}{the number of CPUE index observations used in model fitting}
\item{CPUEobs}{a data frame of CPUE catch rate observations [year, stock, area, subyear, fleet, index] fitted against vulnerable biomass}
\item{nE}{the number of partial F series (typically nfleets)}
\item{nEobs}{the number effort observations (partial F's) one for each catch observations}
\item{Eobs}{a data frame of effort observations one line per observation [year, subyear, area, fleet, effort]}
\item{nCLobs}{the number of length observations}
\item{CLobs}{a data frame of length observations [year, subyear]}
\item{HCobs}{an array of historical catch observations [year x subyear x area x age]}
\item{RAI}{an array of relative abundance - this is pass through data - it is the master index used to derive the standardized effort (partial F data Eobs)}
\item{nI}{the number of fishery independent indices (e.g. 2: a spawning biomass survey in the GOM and MED)}
\item{nIobs}{the number of fishery independent indices}
\item{Iobs}{a data frame of fishery independent relative abundance observations [year, subyear, area, stock, index number, type (biomass/ssb), index]}
\item{nPSAT}{the number of electronic tags of known stock of origin}
\item{PSAT}{a data frame of electronic tag movements [stock, age, subyear, duration til recapture (subyears), from area, to area, number of tags, total from area]}
\item{nPSAT2}{the number of electronic tags of unknown stock of origin}
\item{PSAT2}{a data frame of electronic tag movements [age, subyear, duration til recapture (subyear), from area, to area, prob(stock1), prob(stock2)]}
\item{nTag}{the number of conventional tagging movement observations}
\item{Tag}{a data frame of conventional tagging observations (currently unused)}
\item{nSOOobs}{the number of stock-of-origin observations}
\item{SOOobs}{a data frame of stock-of-origin observations [stock, age, year, subyear, area, number of observations]}
\item{nsel}{the number of selectivity types (not necessarily nf due to selectivity mirroring)}
\item{seltype}{a vector nsel long specifying the type of selectivity (2 = logistic, 3 = Thompson (potentially domed))}
\item{selind}{a vector nf long specifying the correct selectivity for each fleet}
\item{ratiolim}{a vector indicating the ratio of precision to inflection point for the logistic selectivities}
\item{infleclim}{the min/max for the position of selectivity inflection point relative to maximum length for logistic selectivities}
\item{nma}{the number of movement age groups}
\item{ma}{a vector na long specfying the correct movement for each age class}
\item{nMP}{the number of movement parameters that are estimated}
\item{nmovind}{the number of indices for assigning estimated movement parameters to the movement matrices}
\item{movind}{a data frame containing the indexing (correct destination) of each movement parameter [stock, subyear, from area, to area]}
\item{nmov1}{the number of indices for assiging a zero (not an estimated parameter, rows sum to 1 so the first movement is allocated a zero rather than an estimated parameter)}
\item{mov1}{a data frame containing the indexing (correct destination) of a zero value in the movement matrices}
\item{movtype}{the type of movement model (1: Gravity model, 2: Markov model)}
\item{Ilencat}{the categories of length specific vulnerable biomass}
\item{CobsCV}{a vector nf long, the catch observation error (lognormal sd)}
\item{CLCV_num}{the numerator of the normal observation error on lengths}
\item{RDCV}{the recruitment deviation penalty (sigma R)(lognormal sd)}
\item{SSBfit}{the type of SSBfit 1:SSB0 2:SSBnow}
\item{SSBprior}{a vector np long, an optional prior on current spawning biomass}
\item{SSBpriorCV}{the precision of the SSBprior (lognormal sd)}
\item{SSBinc}{a numerical value representing the ratio of SSB from year SSBy2 / SSBy1}
\item{SSBy}{vector of integers representing years over which desired increase in SSB matches SSBinc}
\item{SSBincstock}{the stock that SSBinc refers to}
\item{BSfrac}{a matrix of fractions, stock by season that is the predicted seasonal biomass of Eastern / Western biomass in Western and Eastern areas, respectively}
\item{FCV}{the prior on deviations from the mean F}
\item{movCV}{the prior on deviations from homogenous movement}
\item{selCV}{the prior on selectivity parameters}
\item{SSBincCV}{the prior precision of SSB trend}
\item{BSfracCV}{the prior precision of the Seasonal biomass fraction}
\item{R0diffCV}{the CV on the log ratio of early versus late R0 for the same stock (where applicable)}
\item{nLHw}{the number of likelihood components (for weighting)}
\item{LHw}{a vector nLHw long specifying the relative weight of the various data types 1 catch, 2 cpue, 3 FIindex, 4 Lcomp, 5 SOO, 6 PSAT, 7 PSAT2, 8 RecDev, 9 mov, 10 sel, 11 SRA, 12 SSB, 13 E-W Distr}
\item{muR_ini}{a vector np long specifying an initial value of mean absolute recruitment}
\item{sel_ini}{a matrix of initial values for selectivity [nf x nl]}
\item{selpar_ini}{a matrix of initial value of selectivity parameters [nf x 3]}
\item{lnF_ini}{a vector nCobs long specifying initial value for fishing mortality rates}
\item{ilnRD_ini}{a vector of initial (year 1 disparities in age comp) recruitment deviations (currently unused)}
\item{lnRD_ini}{a vector ny long of recruitment recruitment deviations}
\item{mov_ini}{an array of initial movement [stock x age x subyear x area from x area to]}
\item{qCPUE_ini}{a vector nCPUEq long of initial catchability estimates for CPUE indices}
\item{qI_ini}{a vector nI long specifying initial value for catchabilities of fishery independent indices}
\item{D_ini}{a vector np long specifying initial (comparative) depletion (SSB current / SSB unfished) estimates}
\item{ComplexRD}{an integer specifying whether recruitments can be blocked or estiamted for each year (currently unused)}
\item{ComplexF}{an integer specifying whether an F should be estimated for each catch observation rather than F=qE (currently unused)}
\item{nF}{the number estimated F parameters (usually =nCobs)}
\item{nMPind}{the number of fishery indices used for informing management procedures in subsequent MSE analysis (currently unused)}
\item{MPind}{a data frame of fishery indices used for informing management procedures in subsequent MSE analysis (currently unused)}
\item{MIinv}{a 1 or 0 determining whether the fit should attempt to be MI invariant}
\item{MICV}{numerical value of the CV in additional to regional FMod deviations}
\item{SpatPr}{matrix of Spatial priors}
\item{debug}{a logical (0 or 1) value determing whether the model should be run for a single iteration to check for errors}
\item{verbose}{a logical (0 or 1) value determing whether more complete information should be provided in each iteration of the model estimation}
\item{datacheck}{a unique number for checking that data were read into the model correctly}
\item{CPUEnames}{a character string nCPUEq long, recording the names of CPUE indices}
\item{Inames}{a character string nI long, recording the names of the fishery indepdendent indices}
\item{nMovExc}{integer the number of rows in the movement exclusion data frame}
\item{MovExc}{a data frame of movement exceptions}
\item{Phases}{the phases of the various parameter estimates (1,2,3)}
\item{ET_LHF}{the type of likelihood function used for electronic tagging data (1,2: zero intercept)}
\item{LC_LHF}{the type of likelihood function used for length composition data (1,2: zero intercept)}
}
}
