% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_documentation.R
\docType{data}
\name{OMI_test}
\alias{OMI_test}
\title{Operating model input object for test unit}
\description{
Operating model input object for test unit
}
\author{
Tom Carruthers \email{t.carruthers@oceans.ubc.ca}
}
\keyword{data}

