% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Implementation_models.R
\name{Overage_30}
\alias{Overage_30}
\title{TAC overages of 30 per cent are simulated up to a maximum harvest rate of 90per cent (an implementation error function of class IE)}
\usage{
Overage_30(U, over = 0.3, maxU = 0.9)
}
\arguments{
\item{U}{A numeric value, vector or array of harvest rates}
}
\value{
an object of dimensions U with a harvest rate cap \code{U, over, maxU}
}
\description{
TAC overages of 30 per cent are simulated up to a maximum harvest rate of 90per cent (an implementation error function of class IE)
}
\examples{
Overage_30((1:10)/10)
}
