% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_documentation.R
\docType{data}
\name{Perfect_Obs}
\alias{Perfect_Obs}
\title{#' Generic perfect information observation error model (for everthing but indices)}
\description{
#' Generic perfect information observation error model (for everthing but indices)
}
\author{
Tom Carruthers \email{t.carruthers@oceans.ubc.ca}
}
\keyword{data}
