% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M3_tools.R
\name{pin_from_par}
\alias{pin_from_par}
\title{Create an M3 parameter initialization file (.pin) from the MLE estimates (.par) of a previous run}
\usage{
pin_from_par(OMdir = "C:/M3", pinfile = NA)
}
\arguments{
\item{OMdir}{a character string representing the location of an M3 par file}
}
\value{
wrotes an M3.pin file in \code{OMdir}
}
\description{
Create an M3 parameter initialization file (.pin) from the MLE estimates (.par) of a previous run
}
\examples{
#pin_from_par("C:/ABT-MSE/M3")
}

