% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSE_functions.R
\name{plot_CatchComp}
\alias{plot_CatchComp}
\title{Plot the catch composition during projections r}
\usage{
plot_CatchComp(MSEobj)
}
\arguments{
\item{MSEobj}{An object of class 'MSE'}

\item{MPs}{A numeric vector of MPs (order is MPs slot of MSE object)}
}
\description{
Plot the catch composition during projections r
}
