% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSE_functions.R
\name{plot_Indices}
\alias{plot_Indices}
\title{Plot the true and simulated indices for an MSE run}
\usage{
plot_Indices(MSEobj, index = 1)
}
\arguments{
\item{MSEobj}{An object of class 'MSE'}

\item{index}{An integer number of the index to be plotted}
}
\description{
Plot the true and simulated indices for an MSE run
}
