% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSE_functions.R
\name{plot_Recruitment}
\alias{plot_Recruitment}
\title{Plot the unfished recruitment, mean and stochastic recruitment of a projection}
\usage{
plot_Recruitment(MSEobj, MPs = 1:2)
}
\arguments{
\item{MSEobj}{An object of class 'MSE'}

\item{MPs}{A numeric vector of MPs (order is MPs slot of MSE object)}
}
\description{
Plot the unfished recruitment, mean and stochastic recruitment of a projection
}
