% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Performance_metrics.R
\name{POF}
\alias{POF}
\title{Probability of Over-Fishing (F>FMSY) over 30 projected years (a performance metrics function of class PM)}
\usage{
POF(MSE, pp = 1)
}
\arguments{
\item{MSE}{An object of class MSE}
}
\value{
a matrix with n Management procedures (MSE@nMP) and nsim (MSE@nsim) columns from \code{MSE}
}
\description{
Probability of Over-Fishing (F>FMSY) over 30 projected years (a performance metrics function of class PM)
}
\examples{
loadABT()
POF(MSE_example)
}
