% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Performance_metrics.R
\docType{data}
\name{POFed}
\alias{POFed}
\title{Probability of Over-Fished status (B<BMSY) over all projected years (a performance metrics function of class ABT_PM)}
\format{An object of class \code{ABT_PM} of length 1.}
\usage{
POFed
}
\arguments{
\item{MSE}{An object of class MSE}
}
\value{
a matrix with n Management procedures (MSE@nMP) and nsim (MSE@nsim) columns from \code{MSE}
}
\description{
Probability of Over-Fished status (B<BMSY) over all projected years (a performance metrics function of class ABT_PM)
}
\examples{
loadABT()
POFed(MSE_example)
}
\keyword{datasets}

