% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Performance_metrics.R
\name{POS}
\alias{POS}
\title{Probability of Over-Fished status (B<BMSY) after 30 projected years (a performance metrics function of class PM)}
\usage{
POS(MSE, pp = 1)
}
\arguments{
\item{MSE}{An object of class MSE}
}
\value{
a matrix with n Management procedures (MSE@nMP) and nsim (MSE@nsim) columns from \code{MSE}
}
\description{
Probability of Over-Fished status (B<BMSY) after 30 projected years (a performance metrics function of class PM)
}
\examples{
loadABT()
POS(MSE_example)
}
