% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCRs.R
\name{powdif}
\alias{powdif}
\title{Power difference function.}
\usage{
powdif(x, z, g)
}
\arguments{
\item{x}{Real number: the absolute difference between two numbers}

\item{z}{Real number: the exponent the difference will be raised to}

\item{g}{Real number: the gradient in the exponential difference}
}
\value{
a positive real number
}
\description{
Power difference function.
}
\examples{
powdif(-2,3,1)
}
