% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSE_functions.R
\name{PPlot}
\alias{PPlot}
\title{Performance plot}
\usage{
PPlot(
  MSE,
  Pnames = c("AvC30", "C10", "C30", "D30", "LD", "DNC", "LDNC", "PGK", "AAVC", "Br30")
)
}
\arguments{
\item{MSE}{An MSE object}
}
\value{
a plot showing performance metric statistics accross MPs
}
\description{
Performance plot
}
\examples{
PPlot()
}
