% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Performance_metrics.R
\name{RD30}
\alias{RD30}
\title{Relative depletion (spawning biomass relative to dynamic B0) after 30 years (a performance metrics function of class PM)}
\usage{
RD30(MSE, pp = 1)
}
\arguments{
\item{MSE}{An object of class MSE}
}
\value{
a matrix with n Management procedures (MSE@nMP) and nsim (MSE@nsim) columns from \code{MSE}
}
\description{
Relative depletion (spawning biomass relative to dynamic B0) after 30 years (a performance metrics function of class PM)
}
\examples{
loadABT()
RD30(MSE_example)
}
