% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M3_tools.R
\name{read.fit}
\alias{read.fit}
\title{Read the outputs of an ADMB model run}
\usage{
read.fit(file = "C:/M3", digits = 12, cor_ignore = T)
}
\arguments{
\item{file}{a character string representing the location of ADMB model outputs (.par, .cor, .rep)}

\item{digits}{how many decimal places for numbers}
}
\value{
a list of model parameter estimates and a hessian matrix in folder \code{file}
}
\description{
Read the outputs of an ADMB model run
}
\examples{
#read.fit("C:/ABT-MSE/M3")
}
\author{
this is code that was originally written by Anders Nielsen
}
