% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M3_tools.R
\name{read.hessian}
\alias{read.hessian}
\title{Plot area definitions of an OMI, OM or MSE object}
\usage{
read.hessian(file = "C:/M3")
}
\arguments{
\item{.Object}{an object of class OMI, OM or MSE}
}
\value{
a map of the areas of \code{classy}
}
\description{
Plot area definitions of an OMI, OM or MSE object
}
\examples{
loadABT()
areaplot(OM_1)
areaplot(OMI_1)
areaplot(MSE_example)
}
