% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSE_functions.R
\name{Results_compiler}
\alias{Results_compiler}
\title{Compile the performance results for a complete set of MSE runs (all reference and robustness operating models)}
\usage{
Results_compiler(dir = NULL, name = NULL, CMPdesc = NULL, CMPcode = NULL)
}
\arguments{
\item{dir}{Character string, a directory where the MSE objects are saved (e.g. MSE_1.rda, MSE_2.rda, ... MSE_R_1.rda). Defaults to the current working directory}

\item{name}{character string, the name of your set of MSE runs. Defaults to NULL}

\item{CMPdesc}{A vector of character strings. These describe the CMPs tested in the MSE runs (not including the zero catch CMP that is run by default). Defaults to NULL.}

\item{CMPcode}{A vector of 3 character strings that identify your CMP for presentation purposes (e.g. c("TC1","TC2","TC3","TC4"))}
}
\description{
Compile the performance results for a complete set of MSE runs (all reference and robustness operating models)
}
