% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSE_functions.R
\name{runAll}
\alias{runAll}
\title{Run a set of MSE runs}
\usage{
runAll(OMdirs, MPs, Obs, outfolder)
}
\arguments{
\item{MPs}{A vector of management procedures for analysis}

\item{OMdir}{A vector of operating model directories (e.g. those found in abft-mse/Objects/OMs/)}

\item{nsim}{The number of simulations}

\item{OMdir}{Obs an observation error model of class 'Obs' e.g. Perfect_Obs}

\item{rebuildOMs}{Logical: should the OMs be rebuilt or should those in the directory be used directly}
}
\description{
Run a set of MSE runs
}
