% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M3_tools.R
\name{runM3}
\alias{runM3}
\title{Run an M3 operating model (a multistock spatial seasonal statistical catch at length model)}
\usage{
runM3(OMdir = "C:/ABT-MSE/M3", hess = F, mcmc = F, nits = 10000,
  thin = 40)
}
\arguments{
\item{OMdir}{a character string representing the location of an M3 folder containing M3.exe, M3.dat and stats.cxx files}
}
\value{
runs the M3 model to create standardized ADMB reporting in folder \code{OMdir}
}
\description{
Run an M3 operating model (a multistock spatial seasonal statistical catch at length model)
}
\examples{
#runM3("C:/ABT-MSE/M3")
}
