% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M3_tools.R
\name{runM3p}
\alias{runM3p}
\title{Run an M3 operating model (a multistock spatial seasonal statistical catch at length model)}
\usage{
runM3p(x, OMdir = "C:/M3temp", hess = F)
}
\arguments{
\item{x}{the position in the vector OMdir (one of the folders containing the M3 exe)}

\item{OMdir}{a character string representing the location of an M3 folder containing M3.exe, M3.dat and stats.cxx files}
}
\value{
runs the M3 model to create standardized ADMB reporting in folder \code{OMdir}
}
\description{
Run an M3 operating model (a multistock spatial seasonal statistical catch at length model)
}
\examples{
#runM3p(1,"C:/ABT-MSE/M3")
}
