% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal.R
\name{sdconv}
\alias{sdconv}
\title{Calculate the lognormal standard deviation from logspace mean and standard deviations}
\usage{
sdconv(m, sd)
}
\arguments{
\item{m}{a positive real number}

\item{sd}{a positive real number}
}
\value{
the lognormal standard deviation corresponding to \code{m, sd}
}
\description{
Calculate the lognormal standard deviation from logspace mean and standard deviations
}
\examples{
sdconv(0.5,0.2)
}

