% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunShiny.R
\name{Shiny}
\alias{Shiny}
\title{Run a Shiny Application}
\usage{
Shiny(app)
}
\arguments{
\item{app}{The name of the Shiny application to run.}
}
\description{
\code{Shiny} runs one of the Shiny Applications that are included in the package
}
\references{
Modified from Deal Attali's code: \url{http://deanattali.com/2015/04/21/r-package-shiny-app/}
}
