% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/External.R
\name{sourceall}
\alias{sourceall}
\title{Load all ABT-MSE R code including interal functions}
\usage{
sourceall(ABTdir = "C:/ABT-MSE/")
}
\value{
all functions are available in the global namespace \code{}
}
\description{
Load all ABT-MSE R code including interal functions
}
\examples{
sourceall("C:/ABT-MSE/")
}
