% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\name{SP_i7}
\alias{SP_i7}
\title{A 3 parameter surplus production assessmetn using INDEX 7 (GOM_LAR_SUV) (observation error only)(a management procedure of class MP).}
\usage{
SP_i7(x, dset)
}
\arguments{
\item{x}{a simulation number.}

\item{dset}{a list of simulated data for use by management procedures.}
}
\value{
a TAC recommendation arising from \code{x, dset}.
}
\description{
A 3 parameter surplus production assessmetn using INDEX 7 (GOM_LAR_SUV) (observation error only)(a management procedure of class MP).
}
\examples{
SP_i7(1,dset_example_East)
sapply(1:10,SP_i7,dset_example_East)
}
