% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\docType{data}
\name{SPslope}
\alias{SPslope}
\title{A surplus production management procedure that attempts to identify the slope of changes in surplus production and biomass to locate a gradient of zero (a management procedure of class ABT_MP).}
\format{An object of class \code{ABT_MP} of length 1.}
\usage{
SPslope
}
\arguments{
\item{x}{a simulation number.}

\item{pset}{a list of simulated data for use by management procedures.}

\item{yrsmth}{the number of years over which to quantify slope in surplus production with biomass}

\item{alp}{the SP gradient limits that lead to TAC changes}

\item{bet}{the degree of responsiveness of the MP}
}
\value{
a TAC recommendation arising from \code{x, pset, yrsmth, alp, bet}.
}
\description{
A surplus production management procedure that attempts to identify the slope of changes in surplus production and biomass to locate a gradient of zero (a management procedure of class ABT_MP).
}
\examples{
SPslope(1,pset_example)
sapply(1:10,SPslope,pset_example)
}
\keyword{datasets}

