% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\docType{data}
\name{SPslope_i4}
\alias{SPslope_i4}
\title{A surplus production management procedure using INDEX 2 that attempts to identify the slope of changes in surplus production and biomass to locate a gradient of zero (a management procedure of class MP).}
\format{An object of class \code{MP} of length 1.}
\usage{
SPslope_i4
}
\arguments{
\item{x}{a simulation number.}

\item{dset}{a list of simulated data for use by management procedures.}

\item{yrsmth}{the number of years over which to quantify slope in surplus production with biomass}

\item{alp}{the SP gradient limits that lead to TAC changes}

\item{bet}{the degree of responsiveness of the MP}
}
\value{
a TAC recommendation arising from \code{x, dset}.
}
\description{
A surplus production management procedure using INDEX 2 that attempts to identify the slope of changes in surplus production and biomass to locate a gradient of zero (a management procedure of class MP).
}
\examples{
SPslope_i4(1,dset_example_East)
sapply(1:10,SPslope_i4,dset_example_East)
}
\keyword{datasets}

