% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSE_functions.R
\name{Sub_Results}
\alias{Sub_Results}
\title{Compile the performance results for a complete set of MSE runs (all reference and robustness operating models)}
\usage{
Sub_Results(ResAll, ind)
}
\arguments{
\item{ResAll}{A compiled results object created by Results_compiler() or Join_Results()}

\item{ind}{A vector as long as the MPs (e.g. length(ResAll$MPnames)) or an integer vector (indicating the MPs to keep)}
}
\description{
Compile the performance results for a complete set of MSE runs (all reference and robustness operating models)
}
