% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal.R
\name{tomt}
\alias{tomt}
\title{Transpose an array}
\usage{
tomt(arr)
}
\arguments{
\item{arr}{An array}
}
\value{
the transposition of\code{arr}
}
\description{
Transpose an array
}
\examples{
tomt(array(1:120,c(4,5,6)))
}
