% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSE_functions.R
\name{tradeoff}
\alias{tradeoff}
\title{A generic tradeoff plot based on two functions of class PM}
\usage{
tradeoff(MSE, PMs)
}
\arguments{
\item{MSE}{an object of class MSE}

\item{PMs}{a list of named functions of class PM}
}
\value{
a trade-off figure \code{classy}
}
\description{
A generic tradeoff plot based on two functions of class PM
}
\examples{
loadABT()
tradeoff(MSE_example,list(POF=POF,Y10=Y10))
}

