% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal.R
\name{trlnorm}
\alias{trlnorm}
\title{A log-normal probabilty density function specified according to the transformed mean and standar deviation}
\usage{
trlnorm(reps, mu, cv)
}
\arguments{
\item{reps}{an integer value representing the number of samples}

\item{mu}{the desired mean of the random variable}

\item{cv}{the desired coefficient of variation (StDev/mean) of the random variable}
}
\value{
a random sample produced by \code{reps, mu, cv}
}
\description{
A log-normal probabilty density function specified according to the transformed mean and standar deviation
}
\examples{
x<-trlnorm(1000,2,0.1)
mean(x)
sd(x)/mean(x)
}
