% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\name{U10}
\alias{U10}
\title{Fish at 10 percent harvest rate  (a management procedure of class MP).}
\usage{
U10(x, dset, Urat = 0.1)
}
\arguments{
\item{x}{a simulation number.}

\item{dset}{a list of simulated data for use by management procedures.}
}
\value{
a TAC recommendation arising from \code{x, dset}.
}
\description{
Fish at 10 percent harvest rate  (a management procedure of class MP).
}
\examples{
U10(1,dset_example_West)
sapply(1:10,U10,dset_example_West)
}
