% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\name{U20}
\alias{U20}
\title{Fish at 15harvest rate with imperfect information regarding UMSY and current biomass (a management procedure of class MP).}
\usage{
U20(x, dset, Urat = 0.2)
}
\arguments{
\item{x}{a simulation number.}

\item{dset}{a list of simulated data for use by management procedures.}
}
\value{
a TAC recommendation arising from \code{x, dset}.
}
\description{
Fish at 15harvest rate with imperfect information regarding UMSY and current biomass (a management procedure of class MP).
}
\examples{
U20(1,dset_example_West)
sapply(1:10,U20,dset_example_West)
}
