% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\name{U3}
\alias{U3}
\title{Fish at 3 percent harvest rate given current biomass (a management procedure of class MP).}
\usage{
U3(x, dset, Urat = 0.03)
}
\arguments{
\item{x}{a simulation number.}

\item{dset}{a list of simulated data for use by management procedures.}
}
\value{
a TAC recommendation arising from \code{x, dset}.
}
\description{
Fish at 3 percent harvest rate given current biomass (a management procedure of class MP).
}
\examples{
U3(1,dset_example_West)
sapply(1:10,U5,dset_example_West)
}
