% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Implementation_models.R
\docType{data}
\name{Umax_90}
\alias{Umax_90}
\title{TAC has a cap of 90per cent harvest rate (an implementation error function of class IE)}
\format{An object of class \code{IE} of length 1.}
\usage{
Umax_90
}
\arguments{
\item{U}{A numeric value, vector or array of harvest rates}
}
\value{
an object of dimensions U with a harvest rate cap \code{U, maxU}
}
\description{
TAC has a cap of 90per cent harvest rate (an implementation error function of class IE)
}
\examples{
Umax_90((1:10)/10)
}
\keyword{datasets}

