% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\name{UMSY}
\alias{UMSY}
\title{Fish at MSY harvest rate with imperfect information regarding UMSY and current biomass (a management procedure of class MP).}
\usage{
UMSY(x, dset, Urat = 1)
}
\arguments{
\item{x}{a simulation number.}

\item{dset}{a list of simulated data for use by management procedures.}
}
\value{
a TAC recommendation arising from \code{x, dset}.
}
\description{
Fish at MSY harvest rate with imperfect information regarding UMSY and current biomass (a management procedure of class MP).
}
\examples{
UMSY(1,dset_example_West)
sapply(1:10,UMSY,dset_example_West)
}
