% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\docType{data}
\name{UMSY_PI}
\alias{UMSY_PI}
\title{Fish at MSY harvest rate with perfect information of UMSY and current biomass (a management procedure of class MP).}
\format{An object of class \code{MP} of length 1.}
\usage{
UMSY_PI
}
\arguments{
\item{x}{a simulation number.}

\item{dset}{a list of simulated data for use by management procedures.}
}
\value{
a TAC recommendation arising from \code{x, dset}.
}
\description{
Fish at MSY harvest rate with perfect information of UMSY and current biomass (a management procedure of class MP).
}
\examples{
UMSY_PI(1,dset_example_West)
sapply(1:10,UMSY_PI,dset_example_West)
}
\keyword{datasets}

