% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\name{XSA}
\alias{XSA}
\title{Extended Survival Analysis MP (currently in development)}
\usage{
XSA(x, dset)
}
\arguments{
\item{x}{a simulation number.}

\item{dset}{a list of simulated data for use by management procedures.}
}
\value{
a TAC recommendation arising from \code{x, dset}.
}
\description{
Extended Survival Analysis MP (currently in development)
}
\examples{
XSA(1,dset_example_East)
sapply(1:10,XSA,dset_example_East)
}
