% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Performance_metrics.R
\docType{data}
\name{Y20}
\alias{Y20}
\title{Yield over projected years 11-20 (a performance metrics function of class PM)}
\format{An object of class \code{PM} of length 1.}
\usage{
Y20
}
\arguments{
\item{MSE}{An object of class MSE}
}
\value{
a matrix with n Management procedures (MSE@nMP) and nsim (MSE@nsim) columns from \code{MSE}
}
\description{
Yield over projected years 11-20 (a performance metrics function of class PM)
}
\examples{
loadABT()
Y20(MSE_example)
}
\keyword{datasets}

