% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\docType{data}
\name{ZeroC}
\alias{ZeroC}
\title{No catches (actually very small catches) (a management procedure of class ABT_MP).}
\format{An object of class \code{ABT_MP} of length 1.}
\usage{
ZeroC
}
\arguments{
\item{x}{a simulation number.}

\item{pset}{a list of simulated data for use by management procedures.}
}
\value{
a TAC recommendation arising from \code{x, pset}.
}
\description{
No catches (actually very small catches) (a management procedure of class ABT_MP).
}
\examples{
ZeroC(1,pset_example)
sapply(1:10,ZeroC,pset_example)
}
\keyword{datasets}

