% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Performance_metrics.R
\docType{data}
\name{AAVY}
\alias{AAVY}
\title{Average Annual Variability in Yield over the first 30 projection years (a performance metrics function of class PM)}
\format{An object of class \code{PM} of length 1.}
\usage{
AAVY
}
\arguments{
\item{MSE}{An object of class MSE}
}
\value{
a matrix with n Management procedures (MSE@nMP) and nsim (MSE@nsim) columns from \code{MSE}
}
\description{
Average Annual Variability in Yield over the first 30 projection years (a performance metrics function of class PM)
}
\examples{
loadABT()
AAVY(MSE_example)
}
\keyword{datasets}

