% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Performance_metrics.R
\name{D20}
\alias{D20}
\title{Depletion (spawning biomass relative to dynamic B0) after projection year 20 (a performance metrics function of class PM)}
\usage{
D20(MSE, pp = 1)
}
\arguments{
\item{MSE}{An object of class MSE}
}
\value{
a matrix with n Management procedures (MSE@nMP) and nsim (MSE@nsim) columns from \code{MSE}
}
\description{
Depletion (spawning biomass relative to dynamic B0) after projection year 20 (a performance metrics function of class PM)
}
\examples{
loadABT()
D20(MSE_example)
}
