% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Performance_metrics.R
\docType{data}
\name{D30}
\alias{D30}
\title{Depletion (SSB relative to unfished SSB given current recruitment dynamics) for projected years 21-30 (a performance metrics function of class ABT_PM)}
\format{An object of class \code{ABT_PM} of length 1.}
\usage{
D30
}
\arguments{
\item{MSE}{An object of class MSE}
}
\value{
a matrix with n Management procedures (MSE@nMP) and nsim (MSE@nsim) columns from \code{MSE}
}
\description{
Depletion (SSB relative to unfished SSB given current recruitment dynamics) for projected years 21-30 (a performance metrics function of class ABT_PM)
}
\examples{
loadABT()
D30(MSE_example)
}
\keyword{datasets}

