% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Performance_metrics.R
\name{LD}
\alias{LD}
\title{Depletion (spawning biomass relative to dynamic B0) all projected years (a performance metrics function of class PM)}
\usage{
LD(MSE, pp = 1)
}
\arguments{
\item{MSE}{An object of class MSE}
}
\value{
a matrix with n Management procedures (MSE@nMP) and nsim (MSE@nsim) columns from \code{MSE}
}
\description{
Depletion (spawning biomass relative to dynamic B0) all projected years (a performance metrics function of class PM)
}
\examples{
loadABT()
LD(MSE_example)
}
