% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Performance_metrics.R
\docType{data}
\name{LD}
\alias{LD}
\title{Depletion (SSB relative to unfished SSB given current recruitment dynamics) all projected years (a performance metrics function of class ABT_PM)}
\format{An object of class \code{ABT_PM} of length 1.}
\usage{
LD
}
\arguments{
\item{MSE}{An object of class MSE}
}
\value{
a matrix with n Management procedures (MSE@nMP) and nsim (MSE@nsim) columns from \code{MSE}
}
\description{
Depletion (SSB relative to unfished SSB given current recruitment dynamics) all projected years (a performance metrics function of class ABT_PM)
}
\examples{
loadABT()
LD(MSE_example)
}
\keyword{datasets}

