% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\name{LstepCC4}
\alias{LstepCC4}
\title{A length based management procedure that incrementally changes TAC based on recent and long-term length observations (a management procedure of class MP).}
\usage{
LstepCC4(x, dset, yrsmth = 5, xx = 0.3, stepsz = 0.05,
  llim = c(0.96, 0.98, 1.05))
}
\arguments{
\item{x}{a simulation number.}

\item{dset}{a list of simulated data for use by management procedures.}

\item{yrsmth}{the number of years over which to compare length observations}

\item{xx}{alters target catch as a fraction of historical catch TACstar = (1-xx)muC}

\item{stepsz}{the maximum fractional change in TAC between MP updates}

\item{llim}{a vector of three ratios that determine whether the TAC should be reduced by 2 steps, 1 steps and increased by 1 step}
}
\value{
a TAC recommendation arising from \code{x, dset}.
}
\description{
A length based management procedure that incrementally changes TAC based on recent and long-term length observations (a management procedure of class MP).
}
\examples{
LstepCC4(1,dset_example_East)
sapply(1:10,LstepCC4,dset_example_East)
}
