% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\docType{data}
\name{MCD_i4}
\alias{MCD_i4}
\title{A simple MP that uses an index (INDEX 2) to derive depletion then fishing at average catch x depletion x 2(a management procedure of class MP).}
\format{An object of class \code{MP} of length 1.}
\usage{
MCD_i4
}
\arguments{
\item{x}{a simulation number.}

\item{dset}{a list of simulated data for use by management procedures.}
}
\value{
a TAC recommendation arising from \code{x, dset}.
}
\description{
A simple MP that uses an index (INDEX 2) to derive depletion then fishing at average catch x depletion x 2(a management procedure of class MP).
}
\examples{
MCD_i2(1,dset_example_East)
sapply(1:10,MCD_i2,dset_example_East)
}
\keyword{datasets}

