% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Performance_metrics.R
\name{PGK}
\alias{PGK}
\title{Probability of being in the Green Kobe region (F<FMSY AND B>BMSY) over 30 projected years (a performance metrics function of class PM)}
\usage{
PGK(MSE, pp = 1)
}
\arguments{
\item{MSE}{An object of class MSE}
}
\value{
a matrix with n Management procedures (MSE@nMP) and nsim (MSE@nsim) columns from \code{MSE}
}
\description{
Probability of being in the Green Kobe region (F<FMSY AND B>BMSY) over 30 projected years (a performance metrics function of class PM)
}
\examples{
loadABT()
PGK(MSE_example)
}
