% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\docType{data}
\name{SBT2}
\alias{SBT2}
\title{One of two Southern Bluefin Tuna management procedures that uses recruitment trajectory to adjust TAC (a management procedure of class ABT_MP).}
\format{An object of class \code{ABT_MP} of length 1.}
\usage{
SBT2
}
\arguments{
\item{x}{a simulation number.}

\item{pset}{a list of simulated data for use by management procedures.}

\item{epsB}{currently unused}

\item{epsR}{resposiveness of the MP to recruitment gradient}

\item{tauR}{the time lag (years) for evaluating trend in recruitment}

\item{tauB}{currently unused}

\item{gamma}{currently unused}
}
\value{
a TAC recommendation arising from \code{x, pset, epsB, epsR, tauR, tauB, gamma}.
}
\description{
One of two Southern Bluefin Tuna management procedures that uses recruitment trajectory to adjust TAC (a management procedure of class ABT_MP).
}
\examples{
SBT2(1,pset_example)
sapply(1:10,SBT2,pset_example)
}
\keyword{datasets}

