% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\name{SBT2}
\alias{SBT2}
\title{One of two Southern Bluefin Tuna management procedures that uses recruitment trajectory to adjust TAC (a management procedure of class MP).}
\usage{
SBT2(x, dset, epsB = 0.25, epsR = 0.75, tauR = 5, tauB = 7, gamma = 1)
}
\arguments{
\item{x}{a simulation number.}

\item{dset}{a list of simulated data for use by management procedures.}

\item{epsB}{currently unused}

\item{epsR}{resposiveness of the MP to recruitment gradient}

\item{tauR}{the time lag (years) for evaluating trend in recruitment}

\item{tauB}{currently unused}

\item{gamma}{currently unused}
}
\value{
a TAC recommendation arising from \code{x, dset, epsB, epsR, tauR, tauB, gamma}.
}
\description{
One of two Southern Bluefin Tuna management procedures that uses recruitment trajectory to adjust TAC (a management procedure of class MP).
}
\examples{
SBT2(1,dset_example_East)
sapply(1:10,SBT2,dset_example_East)
}
