% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Implementation_models.R
\name{Umax_30}
\alias{Umax_30}
\title{TAC has a cap of 30per cent harvest rate (an implementation error function of class IE)}
\usage{
Umax_30(U, maxU = 0.3)
}
\arguments{
\item{U}{A numeric value, vector or array of harvest rates}
}
\value{
an object of dimensions U with a harvest rate cap \code{U, maxU}
}
\description{
TAC has a cap of 30per cent harvest rate (an implementation error function of class IE)
}
\examples{
Umax_30((1:10)/10)
}
