% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\docType{data}
\name{UMSY_PI}
\alias{UMSY_PI}
\title{Fish at MSY harvest rate with perfect information of UMSY and current biomass (a management procedure of class ABT_MP).}
\format{An object of class \code{ABT_MP} of length 1.}
\usage{
UMSY_PI
}
\arguments{
\item{x}{a simulation number.}

\item{pset}{a list of simulated data for use by management procedures.}
}
\value{
a TAC recommendation arising from \code{x, pset}.
}
\description{
Fish at MSY harvest rate with perfect information of UMSY and current biomass (a management procedure of class ABT_MP).
}
\examples{
UMSY_PI(1,pset_example)
sapply(1:10,UMSY_PI,pset_example)
}
\keyword{datasets}

