% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Performance_metrics.R
\docType{data}
\name{Y10}
\alias{Y10}
\title{Yield over the first 10 projected years (a performance metrics function of class PM)}
\format{An object of class \code{PM} of length 1.}
\usage{
Y10
}
\arguments{
\item{MSE}{An object of class MSE}
}
\value{
a matrix with n Management procedures (MSE@nMP) and nsim (MSE@nsim) columns from \code{MSE}
}
\description{
Yield over the first 10 projected years (a performance metrics function of class PM)
}
\examples{
loadABT()
Y10(MSE_example)
}
\keyword{datasets}

