% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\name{ZeroC}
\alias{ZeroC}
\title{No catches (actually very small catches) (a management procedure of class MP).}
\usage{
ZeroC(x, dset)
}
\arguments{
\item{x}{a simulation number.}

\item{dset}{a list of simulated data for use by management procedures.}
}
\value{
a TAC recommendation arising from \code{x, dset}.
}
\description{
No catches (actually very small catches) (a management procedure of class MP).
}
\examples{
ZeroC(1,dset_example_West)
sapply(1:10,ZeroC,dset_example_West)
}
